pyproject.toml
[project]
name = "tomorrows-news"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.14",
    "alembic>=1.16.4",
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.4",
    "fastapi[standard]>=0.116.1",
    "feedparser>=6.0.11",
    "langchain>=0.3.26",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.3",
    "langgraph-checkpoint-postgres>=2.0.23",
    "lxml-html-clean>=0.4.2",
    "markdownify>=1.1.0",
    "newspaper4k>=0.9.3.1",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary,pool]>=3.2.9",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    # "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "requests>=2.28.0",
    "sqlalchemy>=2.0.41",
]

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.format]
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "ipython>=9.4.0",
]
